requirements.txt
ultralytics>=8.1.0
opencv-python-headless>=4.9.0
paho-mqtt>=2.0.0
python-dotenv>=1.0.0
